reworked debounced file watcher

+++ src-tauri/src/database/schema.rs
@@ -21,7 +21,7 @@ fn create_tables_if_missing(conn: &Connection) -> Result<()> {
             file_size INTEGER,
             file_extension TEXT,
             media_type TEXT,
-            video_length INTEGER,
+            video_length REAL,
             created_at INTEGER,
             updated_at INTEGER
         )",
